+++ spec/git_spec.sh
@@ -1,71 +1,116 @@
 #!/usr/bin/env bash
 
-# TODO: Optimize usage of temporary directories by using /tmp instead.
+# Set to PWD to ensure that 'git' is the local git wrapper.
+export PATH=$PWD:$PATH
 
-export CONTAINER_NAME=git-command-test
+get_git_command_filetype() {
+  file -b "$(command -v git)"
+}
 
-Describe 'git command'
-  setup() { docker build -t "$CONTAINER_NAME" . &>/dev/null; }
-  cleanup() { docker image rm "$CONTAINER_NAME" &>/dev/null; }
+run_git_tag_version() {
+  local tag="$1"
+  IMAGE_TAG=$tag git --version
+}
 
-  BeforeAll 'setup'
-  AfterAll 'cleanup'
+create_and_clone_into_bare_repo() {
+  local directory="$1"
+  cd "$directory" || exit 1
+  mkdir repo.git
+  cd repo.git || exit 1
+  git init --bare -b main
+  cd .. || exit 1
+  git clone repo.git
+}
 
-  check_git_wrapper() { head -n1 "$(command -v git)" && file "$(command -v git)" | cut -d' ' -f2-; }
-  run_git_version() { git --version; }
-  create_and_clone_into_bare_repo() { cd "$1" && mkdir repo.git && cd repo.git && git init --bare && cd .. && git clone repo.git; }
-  clone_to_relative_path() { tar -xf ./spec/data/test-repo.tar -C "$1" && cd "$1" && git clone test-repo relative-path-repo; }
-  clone_to_absolute_path() { tar -xf ./spec/data/test-repo.tar -C "$1" && cd "$1" && git clone test-repo "$(realpath "$1"/absolute-path-repo)"; }
+clone_to_relative_path() {
+  tar -xf ./spec/data/test-repo.tar -C "$1"
+  cd "$1" || exit 1
+  git clone test-repo relative-path-repo
+}
 
-  # This test ensures that all test cases actually use the Git command wrapper.
-  # Otherwise, we would just use any Git command which might be some other Git
-  # command.
-  It 'Ensures that Git command is actually the Bash wrapper for Git'
-    When call check_git_wrapper
-    The output should eq "#!/usr/bin/env bash
-Bourne-Again shell script, ASCII text executable"
+clone_to_absolute_path() {
+  tar -xf ./spec/data/test-repo.tar -C "$1"
+  cd "$1" || exit 1
+  git clone test-repo "$(realpath "$1"/absolute-path-repo)"
+}
+
+Describe 'git wrapper check'
+  It 'Ensures that git command is the local wrapper'
+    When call command -v git
+    The output should eq "$PWD/git"
+    The status should be success
   End
 
-  # The Git version must match the version that is defined in the Dockerfile.
+  It 'Ensures that Git command is a Bash script'
+    When call file -b "$(command -v git)"
+    The output should eq "Bourne-Again shell script, ASCII text executable"
+    The status should be success
+  End
+
+End
+
+Describe 'git version and help'
+  It 'Prints version information'
+    When call git --version
+    The output should match pattern "git version *.*.*"
+    The status should be success
+  End
+
+  It 'Prints help output'
+    When call git --help
+    The line 1 should include "usage: git"
+    The status should be success
+  End
+End
+
+Describe 'git version can be modified via image tag'
   It 'Prints installed version information of Git'
-    When call run_git_version
-    The output should start with "git version 2.43."
+    When call run_git_tag_version 2.49.1
+    The output should include "git version 2.49.1"
+    The status should be success
   End
 
+  It 'Allows setting specific image tag'
+    When call run_git_tag_version 2.49.0
+    The output should include "git version 2.49.0"
+    The status should be success
+  End
+End
+
+Describe 'git clone'
   # Related issue: https://github.com/jnk22/libreelec-git-command/issues/2
   It 'Creates a bare repo and clones it into another directory'
-    TMPDIR_BARE="$(mktemp -d)"
-    trap 'rm -rf -- "$TMPDIR_BARE"' EXIT
-    Path RepoPath="$TMPDIR_BARE"
-    When call create_and_clone_into_bare_repo "$TMPDIR_BARE"
+    tmpdir_bare="$(mktemp -d)"
+    trap 'rm -rf -- "$tmpdir_bare"' EXIT
+    Path RepoPath="$tmpdir_bare"
+    When call create_and_clone_into_bare_repo "$tmpdir_bare"
     The status should be success
     The path RepoPath should be exist
-    The output should eq "Initialized empty Git repository in $TMPDIR_BARE/repo.git/"
-    The error should eq "Cloning into 'repo'...
-warning: You appear to have cloned an empty repository.
-done."
+    The line 1 should include "Initialized empty Git repository in $tmpdir_bare/repo.git/"
+    The line 2 should include "Cloning into 'repo'..."
+    The line 3 should include "warning: You appear to have cloned an empty repository."
+    The line 4 should include "done."
   End
 
   It 'Clones test repository to a relative path in /tmp directory on host'
-    TMPDIR_RELATIVE="$(mktemp -d)"
-    trap 'rm -rf -- "$TMPDIR_RELATIVE"' EXIT
-    Path RepoPath="$TMPDIR_RELATIVE"/relative-path-repo
-    When call clone_to_relative_path "$TMPDIR_RELATIVE"
-    The status should be success
+    tmpdir_relative="$(mktemp -d)"
+    trap 'rm -rf -- "$tmpdir_relative"' EXIT
+    Path RepoPath="$tmpdir_relative"/relative-path-repo
+    When call clone_to_relative_path "$tmpdir_relative"
+    The line 1 should include "Cloning into 'relative-path-repo'..."
+    The line 2 should include "done."
     The path RepoPath should be exist
-    The error should eq "Cloning into 'relative-path-repo'...
-done."
+    The status should be success
   End
 
-  It 'Clones a repository to absolute path in /tmp directory on host'
-    TMPDIR_ABSOLUTE="$(mktemp -d)"
-    trap 'rm -rf -- "$TMPDIR_ABSOLUTE"' EXIT
-    Path RepoPath="$TMPDIR_ABSOLUTE"/absolute-path-repo
-    When call clone_to_absolute_path "$TMPDIR_ABSOLUTE"
-    The status should be success
+  It 'Clones test repository to an absolute path in /tmp directory on host'
+    tmpdir_absolute="$(mktemp -d)"
+    trap 'rm -rf -- "$tmpdir_absolute"' EXIT
+    Path RepoPath="$tmpdir_absolute"/absolute-path-repo
+    When call clone_to_absolute_path "$tmpdir_absolute"
+    The line 1 should include "Cloning into '$tmpdir_absolute/absolute-path-repo'..."
+    The line 2 should include "done."
     The path RepoPath should be exist
-    The error should eq "Cloning into '$TMPDIR_ABSOLUTE/absolute-path-repo'...
-done."
+    The status should be success
   End
-
 End